Move the push_it command to the repo
The push_it command was made to create the commits to a git repo with
GPG signature forced and uses the current date in UTC for the log,
then push the commit to both the GitHub and GitLab repos. There is no
attempt made to determine if either, or both, repo is missing, and an
error can be expected if used with either repo missing.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>