Add dir-merger script
The dir-merger script, still incomplete, is for moving one directory
tree into another. The intended trees are copies of a source tree
which has varied over time. Common files, which are hash-matched are
skipped while same-named files which are not hash-matched are given a
numeric suffix. The result should be a union of the two, stored in
second argument, with missing directories created as needed.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>