Modify for multiple merges
When a merge is made some files can have an extension added for
resolving a conflicted merge. If the results of one merge is then
merged into another merge result, it is possible for two files to
exist with different suffixes yet have the same content in their
original name.

Comparing files using the base, un-suffixed, name and rebuilding the
suffixes as needed solves the issue, and keeps duplicates from being
formed as a result of the merge. Duplicates within the covered area,
which are not same-named, are still not handled by the process.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>