Change display values for lsync
Switch the display for data sizes from multiples of 1000, to powers of
1024.

Add information to the usage about limiting the function to a single
entry in the directory. Setting traps and tests, and handling globs
is more trouble than I want to add to what is really an alias, and
probably should be anyway.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>