Add output to lsync
For when lsync is used in a script it's nice to know what is being done. To this
point I had been making the script say what was being processed. This switches
the notification from the calling script to lsync itself, and includes the same
information, even without a calling script.

I had also been manually checking the copied results with the original to verify
that the copy operation was successfull. Removed that task by having lsync do a
size comparison between source and copy. Same size should be the same results.
Much faster than trying to determine if the copy was a file or a directory and
trying to walk both versions doing a digest. Less CPU time needed as well.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>