Add verbosity to dir-merger
When processing merges for very large trees, or when the tree has some
large files, such as ISO images, I keep wondering if it's still doing
something or if it's hung up on a bug.

I'm sure that the stream of output can get annoying, but I'm willing
to live with too much output over too little information.

I can always revert the commit if I really don't like the results.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>