Add brains to push_it
The push_it command was blindly trying to push to github and gitlab. Projects
with only one or the other would emit errors, which I was expecting, and would
ignore. Now that I have another host to use I found that using the command to
push to two of them and manually pushing to the third was becomming tiresome.

Expanded the command to lookup all remotes for the projet, filter out those with
the http protocol, and limit it to "push" remotes. For now, my git hosts all use
the git urls, and the http urls are for use when I'm not at the computer, and do
not have the SSH keys to use the git urls. If and when this setup changes I'll
have to revisit this selection process.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>